Instant.now()
@adam

Follow up of the office discussion about:

```
LocalDateTime.now().atZone(ZoneId.systemDefault()).toInstant()
```

java.util.Date doesn't have a time zone attached to it. In fact AWS is not making any choice of a "default time zone" when they return it in their SDK, because java.util.Date is essentially just a Long.

The problem we have is "create an Instant", not "create a datetime in UTC time zone".

So now that we don't have to worry about timezones, what is the most perfect way to create an Instant ?

This change does it perfectly (^_^)