Scala style is 2 spaces, updates editorconfig accordingly
The global editrconfig declaration was set to 4 spaces, which is at
odds with both Scala style and the existing codebase. This sets the
default to 2 spaces for this project.

We might consider adding filetype-specific overrides to this
configuration (e.g. `xml`, which currently uses 4 spaces in the
project), but my focus is on fixing this for the existing bulk of the
program, which is Scala.