Add `--help` option
TLDR; less scrolling required to see the helpful error message when you make a mistake running `ssm`

scopt has a special option with a predefined action called help("help"). When help("help") is defined, parser will print out short error message when it fails instead of printing the entire usage text.

This is useful, because the existing behaviour is to display the short error message and then the long usage text, which masks the helpful error message.