Update the Scala version in the generate executable script
Since this script includes the hard-coded Scala version, it is
essential that this is kept in sync with the Scala version defined in
the build.sbt file. We might consider automating this coupling away,
but for now I've fixed the reference and added a comment to the
build.sbt file so it is less likely these versions drift apart in the
future.