Simplify GitHub Actions workflows
We now use our own setup-scala workflow, instead of having to set up
Java and sbt separately. This requires that we specify a java version
in our `.tool-versions` file.

Note: We're only specifying the major Java version so that we aren't
providing irrelevant minor and patch versions that will drift out of
date over time. This is not supported by `asdf`, but switching to
`mise` (https://mise.jdx.dev/) will work as a drop-in replacement.

We also bump the actions/checkout version.