Remove REPL mode from ssm-scala
This mode of execution was added very early in ssm-scala's life and
has never been properly supported or maintained. A cursory experiment
suggests this feature is currently broken.

The only feature REPL mode offers over and above the standard ssh mode
is the ability to run commands on multiple instances at once, but in
the modern world I wouldn't expect us to be taking this approach to
instance management.